Refactor ContactMe component and add ContactForm
